+++ miden-crypto/src/hash/algebraic_sponge/mod.rs
@@ -17,6 +17,13 @@ use core::ops::Range;
 use super::{Felt, Word, ZERO};
 use crate::field::BasedVectorSpace;
 
+#[cfg(any(
+    all(target_arch = "x86_64", target_feature = "avx2"),
+    all(target_arch = "x86_64", target_feature = "avx512f"),
+    all(target_arch = "aarch64", target_feature = "neon"),
+    all(target_arch = "wasm32", target_feature = "simd128"),
+))]
+mod packed;
 pub(crate) mod poseidon2;
 pub(crate) mod rescue;